crewai-data-scientists/requirements.txt
# Core CrewAI and LangChain
crewai==0.86.0
langchain==0.3.7
langchain-anthropic>=0.2.0
langchain-core==0.3.15

# Web Interface
gradio==4.42.0

# Data Science Libraries
pandas>=2.0.0
numpy<2.0,>=1.26.0
scikit-learn>=1.3.0

# Utilities
python-dotenv==1.0.0

# Compatibility fixes
huggingface_hub==0.23.0
pydantic>=2.4.1,<2.10  # CrewAI 0.86.0 requires Pydantic <2.10
setuptools>=65.0.0
packaging<25,>=23.2
urllib3<2.4,>=1.24.2

# Optional: Enhanced ML models
# xgboost>=2.0.0
# lightgbm>=4.0.0
